refactor/basic: cleancode
