+++ src/main/java/com/sparta/team5/fractal/domain/tag/controller/TagControllerV2.java
@@ -1,8 +1,8 @@
 package com.sparta.team5.fractal.domain.tag.controller;
 
-import com.sparta.team5.fractal.common.annotation.Auth;
-import com.sparta.team5.fractal.common.dto.AuthUser;
-import com.sparta.team5.fractal.common.response.ApiResponse;
+import com.sparta.team5.fractal.common.core.dto.ApiResponse;
+import com.sparta.team5.fractal.common.core.dto.AuthUser;
+import com.sparta.team5.fractal.common.crosscutting.annotation.Auth;
 import com.sparta.team5.fractal.domain.tag.dto.response.TagProductResponse;
 import com.sparta.team5.fractal.domain.tag.dto.response.TagResponse;
 import com.sparta.team5.fractal.domain.tag.service.TagProductServiceV2;
