+++ src/main/java/com/sparta/team5/fractal/domain/tag/service/TagProductServiceV1.java
@@ -1,6 +1,6 @@
 package com.sparta.team5.fractal.domain.tag.service;
 
-import com.sparta.team5.fractal.common.exception.GlobalException;
+import com.sparta.team5.fractal.common.core.exception.GlobalException;
 import com.sparta.team5.fractal.domain.product.dto.ProductSimpleResponse;
 import com.sparta.team5.fractal.domain.product.entity.Product;
 import com.sparta.team5.fractal.domain.product.service.ProductServiceApi;
