+++ src/main/java/com/sparta/team5/fractal/domain/tag/service/TagProductServiceV2.java
@@ -1,7 +1,7 @@
 package com.sparta.team5.fractal.domain.tag.service;
 
-import com.sparta.team5.fractal.common.cache.CacheUtil;
-import com.sparta.team5.fractal.common.exception.GlobalException;
+import com.sparta.team5.fractal.common.core.exception.GlobalException;
+import com.sparta.team5.fractal.common.infra.cache.CacheUtil;
 import com.sparta.team5.fractal.domain.product.dto.ProductSimpleResponse;
 import com.sparta.team5.fractal.domain.tag.dto.response.TagProductResponse;
 import com.sparta.team5.fractal.domain.tag.entity.Tag;
